Initial sealable trie implementation (#1)
As per the documentation, the sealable trie data structure is a Merkle
Patricia Trie with an additional feature where values can be sealed
such that they can no longer be changed but also can no longer be
accessed.  The second property allows the trie to remove nodes keeping
sealed values thus reducing storage requirements for the trie.

As often is the case, there’s room to add more tests to the code, but
for the large part the trie is functional.