Move memory and stdx to separate crates
stdx includes functions which are not dependent on any other part of the
code base and are useful in any code.  As such, it doesn’t make sense
to keep it inside of the sealable-trie crate.  Moving it to separate
crate makes it more reusable.

Separating memory into its own crate should make it easier to implement
pool allocators since they won’t need to link to the sealable-trie crate
and thus ‘know’ the details of RawNode structure.