Move CryptoHash to a new lib crate
With CryptoHash in a separate crate, other crates can be written to
use it without having to rely on sealable-trie.  This should help with
isolation of interfaces, preservation of abstractions and testability
of all the components.

lib crate is also going to be a place for other code which is used
across crates without clear ownership in any of them.