solana-ibc: refactor TrieKey construction
Rather than having TrieKey as an enum which is then converted to
a vector, change it so that it holds the vector directly and upon
creation the key is generated.  This simplifies the usage and avoids
intermediate String allocations.

Also change a little how keys are constructed.  Rather than using a separator
in the generated paths, prefix variable-length values (namely strings) with
their length.