sealable-trie: introduce U3 type and use it for offset
Encode in type that slice’s offset can be at most 7 by introducing
a new lib::u3::U3 type and use it for storing the offset.  As the name
implies, the type is written to hold 3-bit unsigned integers, or in
other words numbers from 0 to 7.

This allows offset check to be removed from Slice and Owned
constructors and since everything is encapsulated the users of the
types need only minimal adjustments.

This actually fixes a bug in Owned::bit where the offset argument
wasn’t validated.  Since the only place which called that function
used offset taken from another slice this never created actual issue
but nonetheless the API is now tighter.