sealable-trie: enforce value in Value node not being sealed (#49)
There’s never a situation where value stored in Value node can be
sealed.  This is because when we seal a key, we seal the entire
subtire.  As such, sealed value implies that the child in the Value
node would be sealed as well.  But that means that the entire Value
node would be sealed and thus wouldn’t be stored.

Encode in type that value in Value may never be sealed.