sealable-trie: propagate decode error rather than panicking
Reduce amount of potential panics by propagating errors from decoding
invalid raw nodes. Those errors should never happen since we write
the raw nodes and thus guarantee valid representation but in case the
code is ever used on data that needs validation it’s safer to detect
and propagate errors.

The crux of the change is in RawNode::decode and Reference::from_raw
methods in nodes.rs file. The rest of the changes is just adapting to
RawNode::decode’s return type change.