sealable-trie: don’t use varint for tag
Using varints for the length of proof isn’t worth it in the end.  The
proof will never be longer than 8192 elements (longest possible key
is 512 bytes long) and even with the additional bit for distinguishing
membership from non-membership proofs 16-bits is enough to encode it.

Using varint to encode the tag saves us at most only one byte at the
cost of having bunch of convoluted code.