sealable-trie: add slightly optimised `set_and_seal` method
Fully optimised set and seal operation is still to be done.  For now
add the `set_and_seal` method to the Trie which removes just the most
trivial duplicate work but really serves more as a placeholder so that
clients can start using the method in anticipation of an optimised
version.