solana-ibc: add function for checking the Trie account (#73)
Extract code which creates AccountTrie from the AccountInfo to
a separate function so that it can be easily reused. Furthermore, add
ownership checks which so far were missing.

Long term I’d like to have a wrapper which can be used inside of
Accounts struct but I couldn’t quite figure out the proper way to
implement it. AccountTrie isn’t quite zero copy (since it makes copy
of the header data). This is left for future improvement.