blockchain: add Verifier trait
Add an Verifier trait rather than a verify method on a Signature.  The
purpose is to allow additional context to be provided when checking
signatures.  This is needed in Solana where signature checking
requires access to per-request data (specifically instructions native
program which is passed with the rest of accounts).