solana-ibc: add  `build-sbf` to GitHub Actions (#88)
Adding ‘cargo build-sbf’ CI step which uses the solana-cli to run the
program in Solana environment.  This enables us to catch any bugs while
building the Solana program.  Ideally, we should run the tests which
requires Anchor CLI and takes quite a bit of time to install and run
the tests.  So for the time being, building the program successfully
using SBF should suffice.