solana-ibc: introduce ConnectionIdx type for connection identification (#105)
IBC connections identifiers are in the form `connection-<number>` and
use sequential numbering.  Take advantage of that by introducing
ConnectionIdx type which only holds the number and storing connection
information in a vector rather than a map.  This is analogous to
change already made for clients.

Issue: https://github.com/ComposableFi/emulated-light-client/issues/35