solana-ibc: make some modules public (#119)
Many of the modules are private which makes them unable to use them
when the project is imported in the relayer. There are some interfaces
which are needed to be imported from this crate. So this PR makes the
required modules public.

Maybe we can split them into seperate crates in the future so that
they can be imported as needed.

* made some modules public so that they can be imported

* uncomment generating a block in deliver method