sealable-trie: add bits::Owned::push and pop_back methods
Also, for the time being, change bits::Owned::unshift to panic if length
of the slice ends up too long.  This is consistant with how Vec handles
things and in all places where unshift was used the caller unwrapped the
value so this isn’t such a terrible design for now.