solana-ibc: limit the maximum length of the TrieKey; use on-stack buffer
Instead of using a Vec to store the TrieKey, limit the maximum length
of the TrieKey and store the value within the structure itself.  Since
we know all possible keys and each key component is fixed-size, we can
easily deduce the longest possible key.  Turns out it’s 22.  That’s
a reasonable size of a buffer to keep within the TrieKey instead of
doing a heap allocation with a Vec.