solana-ibc: implement BorshDeserialize for Event (#127)
So far Event lacked deserialisation derive because (to avoid memory
copies and allocations) it used references. This worked great for
emitting the events but Borsh dosen’t support deserialising such
objects.

Replace CryptoHash and PubKey fields with owned values rather than
references and introduce CowBlock for block fields making Event enum
easily deserialisable.

To be honest I’m not happy with the current state of the type. The
implementation is somewhat noisy with the necessity for CowBlock type,
but it should do for now.