sealable-trie: fix test_owned_concat test
The test was meant to run for length from zero to eight inclusively but due
to the use of `0..8` range run for lengths less than eight instead.  Fix
the range and the test which mishandled length equal to eight case.

While at it, also add test_owned_concat_empty test.