solana-ibc: remove packets storage (#133)
Storing packets on chain would easily pile up since they are just
added and not removed. The previous delivered packets aren’t needed
in the smart contract so there is no point in storing them there. The
packets are required by the relayer which can be fetched by querying
the logs.

* remove packets storage

* fix clippy

* remove cloning of message in deliver method