solana-ibc: update ibc dependency (#128)
ibc 0.48 restructured the modules hierarchy in quite a big way which
results in noisy diff around imports especially as ibc likes to use
quite long paths. To make this somewhat more bearable, introduce ibc
module with re-exports of IBC types we care about.

Note that 0.48.0 fails to build in no_std on older rustc (which is
what is used with build-spf), thus go straight to 0.48.1 version.