sealable-trie: prohibit storing values at key prefixes (#141)
Due to the nature of sealable trie, storing values at a prefix of an
existing key isn’t supported.  To understand why, consider the
following sequence of operations:

* set 0000 → foo
* set 0001 → bar
* seal 0000
* seal 0001
* set 000 → baz

After the second seal operation, node 000 is sealed as well.  This is
what makes sealing a worthwhile pursue allowing us to prune subtries
as the leafs are sealed.

But if 000 is sealed than setting 000 → baz cannot happen even though
the key wasn’t explicitly sealed.  However, if that operation was
performed earlier, it would succeed with the key being able to be
modified.

Similarly, it’s already been codded that sealing a key seals the
entire subtrie.

All of this is somewhat surprising and inconsistent.  To avoid all
that confusion, simply prohibit setting values for prefixes of
existing keys.