solana-ibc: initialise chain in Anchor test (#147)
Most notably, initialise the chain in the Anchor test.  Furthermore,
during the initialisation, set up all the storage accounts (private
storage, trie and chain data) so that init_if_needed in other
instructions is no longer necessary.

This prevents re-initialisation attacks as well as (hopefully)
improves efficiency of the instructions (we’re mostly concerned with
memory usage) as account set up no longer happens in all of the
instructions.

While at it, change some logging in the test so it’s easier to read
the logs.