solana-ibc: introduce separate mock_init_escrow instruction (#148)
Split out mint and escrow initialisation out of mock_deliver and into
mock_init_escrow instruction.  This is done so that the former has
fewer things to initialise and we don’t run into stack or heap memory
issues.

Also, sneak in addition of chain field to the MockDeliver to
demonstrate that this indeed achieves desired result.  Previously,
adding the field into the struct would lead to stack frame overflow.