solana-ibc: use vector-backed maps rather than BTreeMap
In attempt to reduce memory usage and cost of serialising and
deserialising accounts, replace usage of BTreeMap with LinearMap which
is a vector-backed map.

Operations on the map have linear complexity, but since we’re not
doing many of them per instruction, chances are that time saved by not
having to build a tree structure when deserialising will make up for
that.

In fact, for small maps—which is what we’re having—it may be faster
regardless.  In anchor_test_deliver some of the instructions spend
fewer compute units while others use more so this would need further
investigation.