solana-ibc: implement host_timestamp and host_height in terms of guest chain (#149)
Rather than using Solana timestamp and slot number for the host
timestamp and height use that information from guest blockchain’s head
block.  This is because in terms of IBC, the ‘host’ is the guest
blockchain since IBC communication happens between counterparty and
the guest blockchain and not directly with Solana.