solana-ibc: remove dependent arguments from send_packet method (#154)
Currently send_packet takes Packet as an argument.  This means that
caller has to know not only port and channel identifiers on our side
but also provide counterparty information and sequence number.  This,
however, makes no sense.

For counterparty information, it makes no sense because then we fetch
that information ourselves to figure out if it’s correct.  How about,
we just populate that information ourselves instead?

For sequence number this is actually a bug since it causes competing
send packet requests to clash.  If next squence number is 42 and two
users want to send a packet, only one of them will be accepted.
Again, instead of requiring user to send that information, fill it
ourselves.

Change the prototype of send_packet method such that it doesn’t take
counterparty information and sequence number.  Instead make the method
derive that information.

Closes: https://github.com/ComposableFi/emulated-light-client/issues/152

---------

Signed-off-by: Dhruv D Jain <dhruv@iamsizzling.com>
Co-authored-by: Michal Nazarewicz <mina86@mina86.com>