solana-ibc: mix client id into trie values whose key depends on client
When computing trie key for paths which include client id, we’re
stripping out the client type part of the id.  So keys to states of
clients ‘foo-0’ and ‘bar-0’ are the same.

To make sure that someone doesn’t produce a malicious proof which uses
wrong client type, (non-)membership verification algorithm must be
able to check the client id.  Make that possible by mixing in the
client id into the hash stored in the trie.

This way, if the path includes client id, the verification can
calculate the hash from value and client id and check if that was
stored in the trie.