sealable-trie: simplify extend implementation
Using generics for extend implementation may be cool but it made the
code overly complex.  Simplify the bits::concat::Concat trait to
handle concatenation with creation of a new Owned object, replace
Owned::extend method by implementation which takes slice as suffix
parameter and add Owned::push_back for the case of appending a single
bit to the Owned slice.