blockchain: separate BlockHeader out of Block (#162)
Separate BlockHeader out of Block with all data except for the next
epoch.  Instead, header only stores the next epoch commitment
(i.e. hash of the serialised epoch).

Reason for doing this is that serialised epoch is noticeably larger
than just the block.  Meanwhile, there are operations which one may
want to perform without having the next epoch provided.

For instance, proving past block hash can be done by sending a chain
of block headers.  In old code, if that chain crosses epoch boundary,
the epoch would have to be included.  Now, only the epoch hash needs
to be provided.

Part of the motivation is also Solana where instructions are quite
limited in size with a single transaction having to fit in
roughtly 1.2 KB.  If an epoch defines 20 validators, it’s serialised
size is already almost 1 KB.