solana-ibc: use nanosecond precision for block timestamp
The intention was to use nanosecond timestamps in BlockHeader however
due to incorrect assumption about Solana’s Clock Block’s timestamp was
populated with seconds timestamp.

Replace solana-ibc code to properly set the timestamp to nanoseconds
timestamp.

Furthermore, rename the field in BlockHeader to `timestamp_ns` to better
document type of the variable.