solana-ibc: add CPI verification with method to get validators. (#182)
For updating the validators stake, `set_stake` will be called from the
restaking program.  To verify if the call is really being made from
it, we pass the instruction key from which we can verify if the call
is really being made from the expected program (restaking program).

Added a method to fetch the current validators from which we can find
their stake and rewards (when we add it).  This method is useful to
get rewards when the user wants the withdraw the funds.

---------

Signed-off-by: Dhruv D Jain <dhruv@iamsizzling.com>
Co-authored-by: Michal Nazarewicz <mina86@mina86.com>