added generic argument in split_at method (#197)
The `stdx::split_at` method has been changed in commit 8c74a55:
‘write-account: add Solana program for writing data into an account’
by adding type generic argument.  Update call site to reflect that.

This fixes a build failure which wasn’t caught because CI on the
offending commit wasn’t rerun between when the pull request was
initially made and when it was merged much later.