restaking: seperating instructions in a seperate file in TS test (#189)
The instructions are kept in a separate file so that they can be
easily used in client side and scripts.  Only the arguments and
a couple of accounts would need to be provided and other accounts
would be handled by method. The instruction can then be just signed
by the wallet and sent.