solana-ibc: use Solana ed25519 verifier for Tendermint client state (#206)
Take advantage of the support for global state in the allocator and
use it for the verifier state.  We’re using that state in patched
Tendermint client implementation to verify signatures there.

At the moment, because of the limitations of Solana we can verify less
than 10 signatures, but this change at least demonstrates the approach
we can take to have Tendermint client state updates implemented.