solana-ibc: Delegation, setting appropriate stake amount and fetching candidate during withdrawal (#216)
When the restaking program would call the `set_stake` method in solana
ibc, we were setting the stake for the caller address which wouldnt be
the case if the stake is being delegated. So supplying the `set stake`
method with validator key indicating the validator to which the stake is
getting delegated to.

When setting stake, add the existing stake to the new stake and call the
`set_stake` method in `solana_ibc` during deposit. During withdrawal,
perform vice versa.

During withdrawal, fetch the stake from the candidate instead of
validator since the candidate may or may not be part of validator in the
next epoch based on the stake.