solana-trie: refactoring (#224)
Firstly, rename `AccountTrie` to `TrieAccount`.  The former suggests
that it’s a type for a trie which holds accounts or some such.  The
latter better describes that it’s an account that is holding the trie.

Secondly, introduce TrieACcount::from_account_info moving
corresponding code from solana-ibc.  This makes solana-trie more aware
of Solana.

Lastly, split the implementation into multiple files so it’s easier to
comprehend what belongs with what.

This introduces no behaviour changes to the code.