Rename ‘blockchain’ crate to ‘guestchain’
The name ‘blockchain’ is fine when working on code base which deals
with only the guest blockchain.  However, it is a bit confusing when
working in code base which deals with multiple different chains.
‘Blockchain’ becomes completely nondescript.

Rename the crate to ‘guestchain’ so that in external projects it’s
easier to recognise what the crate is for.

This is purely refactoring change.  No behaviour changes are
introduced.