lib: guard Solana support by `solana-program` feature
lib crate conditionally includes solana-program dependency if it’s
built for Solana.  This allows using Solana-provided host functions
for calculating the hash without caller having to care about the
details.

However, the issue is that Cargo doesn’t always take target_os into
consideration when resolving dependencies.  To be perfectly honest I’m
not completely sure of the exact mechanics but the way the code is
written now, when working with this crate in another project I’m
getting the following failures:

    error: failed to select a version for `curve25519-dalek`.
        ... required by package `solana-program v1.17.7`

    versions that meet the requirements `^3.2.1` are: 3.2.1

      previously selected package `curve25519-dalek v3.0.0`
        ... which satisfies dependency `curve25519-dalek = "^3"` of package `ed25519-dalek v1.0.1`

Change the code such that solana-program needs to be explicitly
requested by the user.  This resolves the aforementioned issue while
allowing Solana programs to continue using the optimised code.