restaking: added unbonding period to restaking (#234)
Added 7 days of unbonding period to withdrawals. User has to request for
withdrawal where their receipt token is transferred to escrow until the
unbonding period is over. But during that period, user has an option to
cancel the withdrawal request and get back their receipt token. Once the
unbonding period is over, the `withdraw` method can be called to perform
the withdrawal.

---------

Signed-off-by: Dhruv D Jain <dhruv@iamsizzling.com>
Co-authored-by: Michal Nazarewicz <mina86@mina86.com>