solana-ibc: add more fields to block events (#245)
Currently to get a signature that has been added to a block, relayer
would need to listen on BlockSigned events, look for instruction that
resulted in the event being emitted and then parse the signature out
of the instruction data.  This is rather cumbersome so add signature
field to the BlockSigned event.

Furthermore, make it easier to associate block height with the events
by adding block_height filed to BlockSigned and BlockFinalised events.