cf-guest: change encode_to_vec to return Result (#248)
For compatibility with APIs used in other places, change encode_to_vec
method defined via any_convert macro to return a Result. The method is
still infallible and the error type reflects that but it’s now valid to
run `foo.encode_to_vec().unwrap()`. To not deal with the error case,
`encode` method is provided in addition.