cf-guest: implement misbehaviour handling
Add handling of misbehaviours.  Both detection of headers which imply
a misbehaviour as well as handling of explicit Misbehaviour message.

As part of this, add support for accepting blocks from previous epoch.
Previously, once we received a block which is the last block of an
epoch we would only accept headers from the new epoch.  Now we
continue to accept headers from current epoch.