solana-ibc: update revision number of height to 1 (#253)
Updating revision number of height to 1 from 0 since while decoding, the
revision number field which is omitted while encoding ( since its 0 )
fails when decoding because of missing field. It could be fixed by
having a default value in case of missing field but since it would
require upgrading of cosmos node, we can just use revision number as 1.