solana-ibc: store previous host consensus states (#257)
Storing previous consensus states upto a maximum of 20 where the host
consensus state can be queried at a particular height. We are storing 20
for the time being but we can go upto 40 or 50 max so that we dont
consume a lot of compute units since for `ConnectionOpenAck` takes
almost 1.3M CU out of 1.4M.

We need this because the proof height may not always be same as the
current height.