solana-trie: change header and pointer format  (#264)
There are two changes to the format.  They are backwards-incompatible.

Firstly, add more padding to the header so that it is 72 bytes.  This
way the header has the same size as allocated blocks and indexes at
which blocks are stored are aligned to their size.  Secondly, change
the Ptr value from being byte offset to block index.  I.e. rather than
offset 72 or 144 the Ptr is now 1 or 2.

There are two motivating factors for this change.

Firstly, this change makes Ptr space more ‘dense’.  With Solana
account being at most 10 MiB, the highest Ptr value fits in 18 bits.
Ptr is limited to 30 bits so that leaves 12 bits for further
extensions.  For example, if we ever need to split the trie into
multiple accounts, those bits could be used to store 4-bit account
index and 8-bit bump value.

Secondly, Ptr no longer has alignment requirement.  Therefore it’s not
possible for the caller to provide a misaligned address for the node.
All alignment is handled internally by the allocator.  This removes
potential error vector and potential need to verify incoming Ptr
values.  Values out of bound can of course still be provided but those
are naturally checked when indexing the array.