solana-ibc: add wasm consensus and client state (#255)
The client and consensus state stored on the host are in wrapped in
wasm. So we need to add a new wasm client state and consensus state type
so that we can wrap the and unwrap the states.

The client state is available in `ibc-rs` so we dont need to implement
it ourselves. But the consensus state which is available in `ibc-rs`
doesnt have `timestamp` field which the consensus state in relayer has.
So we need to implement the protobuf for consensus state ourselves. This
is implemented in the wasm crate.

---------

Co-authored-by: Michal Nazarewicz <mina86@mina86.com>