proper exception handling in restaking tests (#273)
In some cases, when we check for failure case in restaking tests, we
would want to fail the test if the callback passes. But before we use to
throw the error in try which was getting ignored in catch. So now, we
are throwing the error outside the try if the callback inside passes.