solana-ibc: provide hash implementations to Tendermint code
Take advantage of new features of ibc-rs which allow passing
HostFunctionsProvider and SHA256 implementation to Tendermint code.

So far, IBC would hard-code native implementation whereas what we want
is calls to Solana syscalls.  With recent changes to ibc-rs it’s now
possible for us to pass our own implementation for the hash
calculation.