solana-ibc: log client state and current block height (#279)
There might be some instances where we need client state at a particular
height mostly for finding the proof height of a packet. But since we are
only storing the latest client state and storing previous states is
expensive, we can just log it and fetch it whenever we want by
traversing the previous logs. The height is also logged to find the
proof height which is `height + 1`. Though it might not be used as much,
but its logged for convenience.

---------

Co-authored-by: Michal Nazarewicz <mina86@mina86.com>