solana-ibc: validation for token mint and fix tests (#295)
The token mint on destination chain is a PDA with seeds as portId,
channelId and hashed denom. But during mint and burn, the account was
not validated if it really was a PDA with the above seeds which could
have enabled users to pass a different token mint.

Fixed tests to reflect the above changes.