restaking: add checking for the Instructions sysvar (#296)
The Instructions sysvar isn’t used directly by the restaking contract
and instead is passed via CPI to the solana-ibc contract. As such,
it’s safe to not verify the account since solana-ibc will perform the
verification. Nonetheless, to introduce another layer of protection,
check that the passed account is what we expect.

In addition, remove instruction account from WithdrawalRequest call
since it was never used.