solana-ibc: Check atleast one of the timeouts are non zero. (#303)
When a packet is being sent, the timeout height and timeout timestamp
can both be set to zero which effectively means that the packet can
never time out. And there is no explicit check to prevent this. And this
packet would never be accepted by cosmos since cosmos has a check that
skips the packet if both the timeouts are zero. And since the packet
never times out, the funds would be stuck forever.