solana-ibc: Fix decimals (#302)
Different tokens have different decimals so for every new token that is
bridged should have the token previously be created by calling
`init_mint` which takes original decimals, effective decimals ( decimals
if the OG decimals is very large ), token name, token symbol, token uri
etc. Right now it can only be created by authority which is the fee
collector.

Since solana only supports u64 tokens, we cannot have large decimals
since that would overflow u64. So when we mint a new token on solana, we
set a fixed decimal for it and if the decimal is less than the original
one we chop the rest of the value. This would reduce the precision but
thats what we can do for now.

The decimal on solana would always be equal or less than the decimal of
the token on the other chain.

---------

Co-authored-by: Blas Rodriguez Irizar <rodrigblas@gmail.com>
Co-authored-by: Michal Nazarewicz <mina86@mina86.com>