cf-guest: fix commitment membership proofs verification
Packet commitments, receipts and acks are stored in the trie as is
rather than being hashed.  This is because they are already 32-byte
hashes and there is no need to hash that again.

However, the membership proof verification didn’t take that into
account and hashed the commitments checking if the hash exists in the
trie.

Fix the implementation so that the commitment value is taken raw and
checked that it exists in the trie.