validator: generate block only if the conditions are passed (#339)
Updating the validator client to only generate block if the conditions
are satisfied. The validations are done before the call is being made to
reduce the unnecessary calls to the program.

Reducing the Compute units to 150k for signing the block since as of now
with 24 validators, a maximum of only 70k CU are consumed. The extra CU
are requested in case the validator set increases.

---------

Co-authored-by: Michal Nazarewicz <mina86@mina86.com>