solana-ibc: Unwrap sol when unescrowing wrapped sol. (#348)
Whenever someone sends SOL to Solana, they get wrapped sol instead of
native SOL which is unusable if the wallet didnt have SOL previously. So
whenever someone send SOL from solana, the escrow account which holds
wsol gets closed and all the lamports are sent to the `mint_authority`
account which is the owner of the escrow account. When we unescrow the
wrapped sol, we debit it from the `mint_authority` account instead of
transferring from the `escrow` account. This will allow users to receive
native SOL on solana instead of wrapped sol. Though when users deposit,
they need to have wrapped SOL which can be a prior instruction to
convert SOL to wSOL.