Update Solana dependencies to 1.18.22
When using Solana 1.18 CLI our tests fail with the following error:

    Access violation in unknown section at address 0x9a of size 1

Because of that we keep the CLI at 1.17 (though this commit upgrades
it to latest 1.17 release).

However, updating the dependency to 1.18 branch works without any
issues so we can do that.  Doing so has the benefit that we can
finally update ahash dependency (Solana 1.17 pins it at 0.8.5) which
means that we can now build with nightly Rust.